refactor: adjust repository pattern to survey_answer database
