+++ src/survey/survey.service.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { SurveyAnswer } from '@root/entities/SurveyAnswer.entity';
-import { Repository } from 'typeorm';
 
 import { SurveyRepository } from './survey.repository';
+import { SurveyAnswerRepository } from './survey_answer.repository';
 
 @Injectable()
 export class SurveyService {
@@ -11,25 +10,13 @@ export class SurveyService {
     @InjectRepository(SurveyRepository)
     private surveyRepository: SurveyRepository,
 
-    @InjectRepository(SurveyAnswer)
-    private surveyAnswerRepository: Repository<SurveyAnswer>,
+    @InjectRepository(SurveyAnswerRepository)
+    private surveyAnswerRepository: SurveyAnswerRepository,
   ) {}
 
   async getSurveyByType(type: string): Promise<any[]> {
     const survey = await this.surveyRepository.getSurveyByType(type);
-    const answers = [];
-    for (const question of survey[0]['surveyQuestions']) {
-      answers.push(
-        await this.surveyAnswerRepository.find({
-          relations: ['question', 'answerType', 'singleChoiceAnswers'],
-          where: {
-            question: {
-              questionId: question.questionId,
-            },
-          },
-        }),
-      );
-    }
+    const answers = await this.surveyAnswerRepository.getSurveyAnswers(survey);
     return answers;
   }
 }
