+++ src/survey/survey_answer.repository.ts
@@ -1,0 +1,23 @@
+import { Survey } from '@root/entities/Survey.entity';
+import { SurveyAnswer } from '@root/entities/SurveyAnswer.entity';
+import { Repository, EntityRepository } from 'typeorm';
+
+@EntityRepository(SurveyAnswer)
+export class SurveyAnswerRepository extends Repository<SurveyAnswer> {
+  async getSurveyAnswers(survey: Survey[]): Promise<SurveyAnswer[]> {
+    const answers = [];
+    for (const question of survey[0]['surveyQuestions']) {
+      answers.push(
+        await this.find({
+          relations: ['question', 'answerType', 'singleChoiceAnswers'],
+          where: {
+            question: {
+              questionId: question.questionId,
+            },
+          },
+        }),
+      );
+    }
+    return answers;
+  }
+}
